feat: 도메인을 실제 행위를 기준으로 분리
