+++ src/main/java/com/nextlevel/kky/lotto/LottoShop.java
@@ -1,0 +1,23 @@
+package com.nextlevel.kky.lotto;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class LottoShop {
+
+    private static final int LOTTO_PRICE = 1000;
+    private final IntegerArrayGenerator integerArrayGenerator;
+
+    public LottoShop(IntegerArrayGenerator integerArrayGenerator) {
+        this.integerArrayGenerator = integerArrayGenerator;
+    }
+
+    public List<Lotto> buyLotto(int amount) {
+        int count = amount / LOTTO_PRICE;
+        List<Lotto> lottoList = new ArrayList<>(amount);
+        for (int i = 0; i < count; i++) {
+            lottoList.add(new Lotto(integerArrayGenerator.getIntegerArray()));
+        }
+        return lottoList;
+    }
+}
