+++ src/test/java/com/nextlevel/kky/lotto/LottoShopTest.java
@@ -6,12 +6,12 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class LottoServiceTest {
+public class LottoShopTest {
 
     @Test
     public void buyLottoTest() {
-        LottoService lottoService = new LottoService(() -> List.of(1, 2, 3, 4, 5, 6));
-        List<Lotto> lottoList = lottoService.buyLotto(14000);
+        LottoShop lottoShop = new LottoShop(() -> List.of(1, 2, 3, 4, 5, 6));
+        List<Lotto> lottoList = lottoShop.buyLotto(14000);
 
         assertThat(lottoList.size()).isEqualTo(14);
         assertThat(lottoList.get(0).getNumbers()).contains(1, 2, 3, 4, 5, 6);
@@ -19,11 +19,11 @@ public class LottoServiceTest {
 
     @Test
     public void calculateWinningStatisticsTest() {
-        LottoService lottoService = new LottoService(() -> List.of(1, 2, 3, 4, 5, 6));
+        LottoShop lottoShop = new LottoShop(() -> List.of(1, 2, 3, 4, 5, 6));
         Lotto winningNumbers = new Lotto(List.of(1, 2, 3, 4, 5, 6));
-        List<Lotto> lottoList = lottoService.buyLotto(5000);
+        List<Lotto> lottoList = lottoShop.buyLotto(5000);
 
-        WinningStatistics winningStatistics = lottoService.calculateWinningStatistics(winningNumbers, lottoList);
+        WinningStatistics winningStatistics = LotteryCommission.calculateWinningStatistics(winningNumbers, lottoList);
         assertThat(winningStatistics.getFirst()).isEqualTo(5);
     }
 }
